Game over screen
Created a game over screen which is activated by a game manager game object. When the player loses all of their health the healthController's Die() method is called which calls the GameManager's GameOver() method if the object the that "died" is the player. GameOver() displays the Game Over Canvas. The game over screen also has a restart button which calls Restart() in the GameManager.